Tetris: Reorganise code to save more instructions.
Inline game state, gameboard, and piece stage.
This has the added advantage of moving the game state to the beginning of memory. The lines cleared count is now at address 0x01.

Change the way that piece_x is used so that the default value 0 represents the center of the board.
This allows an instruction to be saved because piece_x can be cleared to 0 to initialize the piece position.